requirements.txt
chainlit==2.7.2
azure-cosmos==4.9.0
openai==1.102.0
python-dotenv==1.1.1
azure-monitor-opentelemetry==1.7.0
opentelemetry-instrumentation-openai-v2==2.1b0